Fix homebrew after upgrading Mac OS X to El Cap
When upgrading any version of Mac OS X, the permissions on /usr/local
get reset to root. Checking the stat on the file allows this to be
fixed.

ignore_errors on the brew upgrade task means the playbook doesn't file
if brew-cask is already at the current version as it returns an error
state.